ADD delete e alterar

+++ frontend/blog/src/app/feed/feed.component.ts
@@ -4,17 +4,19 @@ import { RouterModule } from '@angular/router';
 import { PostService } from '../service/post.service';
 import { Post } from '../model/Post';
 import { FormsModule, ReactiveFormsModule, FormGroup, FormBuilder, Validators } from '@angular/forms';
+import { MatIconModule } from '@angular/material/icon';
+
 
 @Component({
   selector: 'app-feed',
   standalone: true,
-  imports: [RouterModule, CommonModule, FormsModule, ReactiveFormsModule],
+  imports: [RouterModule, CommonModule, FormsModule, ReactiveFormsModule,MatIconModule],
   templateUrl: './feed.component.html',
   styleUrls: ['./feed.component.scss']
 })
 export class FeedComponent implements OnInit {
   listPost: Post[] = [];
-  postAdd: Post = new Post(0, '', '');
+  postAdd: Post = new Post('', '', 0);
   postForm!: FormGroup;
 
   constructor(private postService: PostService, private fb: FormBuilder) {
@@ -37,13 +39,15 @@ export class FeedComponent implements OnInit {
   publicarMensagem() {
     if (this.postForm.valid) {
       const newPost = new Post(
-        this.listPost.length > 0 ? this.listPost[this.listPost.length - 1].id + 1 : 1,
+        // Deixe o ID como 0 ou undefined; o json-server gerará um novo ID.
         this.postForm.value.nome,
-        this.postForm.value.mensagem
+        this.postForm.value.mensagem,
+        undefined
       );
 
       this.postService.postMensagem(newPost).subscribe((post: Post) => {
         this.listPost.push(post);
+        this.findPosts();  // Recarrega os posts para garantir que tudo está sincronizado.
         this.postForm.reset();
       });
     } else {
@@ -51,9 +55,17 @@ export class FeedComponent implements OnInit {
     }
   }
 
+
+
   remover(postARemover: Post) {
-    this.postService.deletePost(postARemover.id).subscribe(() => {
-      this.listPost = this.listPost.filter(post => post.id !== postARemover.id);
-    });
+    if (postARemover.id !== undefined) {
+      this.postService.deletePost(postARemover.id).subscribe(() => {
+        this.listPost = this.listPost.filter(post => post.id !== postARemover.id);
+        this.findPosts();
+      });
+    } else {
+      console.error('ID do post é indefinido, não é possível remover');
+    }
   }
 }
+
